feat: добавить GlobalExceptionHandler и кастомные исключения

+++ src/main/kotlin/com/example/demo/service/impl/AdminServiceImpl.kt
@@ -2,26 +2,34 @@ package com.example.demo.service.impl
 
 import com.example.demo.database.dao.UserDao
 import com.example.demo.database.entity.Users
-import com.example.demo.model.dto.rq.CreateUserRq
 import  com.example.demo.model.dto.rs.UserDto
 import com.example.demo.model.mapper.UserMapper.toDto
 import com.example.demo.service.AdminService
-import com.example.demo.service.UserService
 import org.springframework.stereotype.Service
+import com.example.demo.exeptions.UserDoesntExistException
+import com.example.demo.model.Role
+import org.springframework.security.core.context.SecurityContextHolder
+
 
 @Service
-class AdminServiceImpl (
+class AdminServiceImpl(
     private val userDao: UserDao
-): AdminService{
+) : AdminService {
+
     override fun getUsers(): List<UserDto> {
         return userDao.findAll().map { it.toDto() }
     }
 
     override fun getUserById(id: Long): UserDto {
-        return userDao.findById(id).orElseThrow().toDto()
+        return userDao.findById(id).orElseThrow {
+            UserDoesntExistException("Пользователя с ID $id не существует")
+        }.toDto()
     }
 
     override fun deleteUserById(id: Long) {
-        return userDao.deleteById(id)
+        if (!userDao.existsById(id)) {
+            throw UserDoesntExistException("Невозможно удалить: пользователь с ID $id не найден")
+        }
+        userDao.deleteById(id)
     }
 }
